Add pretty markdown editor to rulebook form
Gollum is the markdown editor that github uses, very nice!

I also tried to fix the issue of hitting a 500 eror when you dont have
any rulebook to fetch for the current page. It still isnt completely
fixed but since we should always have a rulebook it should not really be
a problem.